funbot/requirements.txt
discord.py
discord.py commands
random module
rquest module
json
os